Delete move/pre-move/post-move hooks, these hooks are no longer needed.
Remove move/pre-move/post-move hook references

Mcollective cleanup: remove methods and references related to move, pre-move and post-move.
