--- cartridges/openshift-origin-cartridge-abstract/abstract/info/hooks/move
@@ -1,72 +1,0 @@
-#!/bin/bash
-
-# Move to a new ip
-
-# Exit on any errors
-set -e
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/apache
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/network
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || [ $# -eq 4 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-idle=false
-if [ "$4" == "--idle" ]
-then
-  idle=true
-fi
-
-
-cartridge_type=$(get_cartridge_name_from_path)
-
-setup_basic_vars
-CART_INFO_DIR=$CARTRIDGE_BASE_PATH/${cartridge_type}/info
-
-observe_setup_app_home
-observe_setup_app_and_git_dirs
-observe_setup_cart_instance_dir
-
-#
-# Find an open localhost IP
-#
-IP=`find_open_ip $uid $uuid`
-echo "export OPENSHIFT_INTERNAL_IP='$IP'" > $APP_HOME/.env/OPENSHIFT_INTERNAL_IP
-if [ -f $CART_INFO_DIR/bin/deploy_httpd_config.sh ]
-then
-  $CART_INFO_DIR/bin/deploy_httpd_config.sh $application $uuid $IP
-fi
-run_as_user "${CARTRIDGE_BASE_PATH}/abstract/info/bin/redeploy_config_dir.sh"
-
-. $APP_HOME/.env/OPENSHIFT_INTERNAL_IP
-
-#
-# Create virtualhost definition for apache
-#
-export CART_INFO_DIR
-if [ -f $CART_INFO_DIR/bin/deploy_httpd_proxy.sh ]
-then
-  $CART_INFO_DIR/bin/deploy_httpd_proxy.sh $application $namespace $uuid $OPENSHIFT_INTERNAL_IP
-else
-  $CARTRIDGE_BASE_PATH/abstract/info/bin/deploy_httpd_proxy.sh $application $namespace $uuid $OPENSHIFT_INTERNAL_IP
-fi
-
-if $idle
-then
-  nohup oo-idler -u $uuid &> /dev/null &
-fi
-
