--- cartridges/openshift-origin-cartridge-cron-1.4/info/hooks/move
@@ -1,1 +1,0 @@
-control-scripts
