--- cartridges/openshift-origin-cartridge-haproxy-1.4/info/hooks/post-move
@@ -1,1 +1,0 @@
-#!/bin/bash
