--- cartridges/openshift-origin-cartridge-haproxy-1.4/info/hooks/pre-move
@@ -1,1 +1,0 @@
-#!/bin/bash
