--- cartridges/openshift-origin-cartridge-jenkins-client-1.4/info/hooks/move
@@ -1,29 +1,0 @@
-#!/bin/bash
-
-# Move to a new ip
-
-# Exit on any errors
-set -e
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/apache
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/network
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-
-setup_basic_vars
