--- cartridges/openshift-origin-cartridge-mongodb-2.2/info/hooks/move
@@ -1,80 +1,0 @@
-#!/bin/bash
-
-# Move to a new ip
-
-# Exit on any errors
-set -e
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/apache
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/network
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-
-cartridge_type="mongodb-2.2"
-setup_basic_vars
-
-CART_INFO_DIR=${CARTRIDGE_BASE_PATH}/embedded/$cartridge_type/info
-CART_ETC_DIR=$CART_INFO_DIR/configuration/etc
-source ${CART_INFO_DIR}/lib/util
-
-MONGODB_DIR="$APP_HOME/$cartridge_type"
-
-observe_setup_var_lib_dir "$MONGODB_DIR"
-
-# This gear is dedicated to running mongodb - configure it as such.
-if only_cart_on_gear $cartridge_type; then
-    prepare_gear_for_standalone_mongodb
-fi 
-
-MONGODB_PORT=27017
-MONGODB_IP=`find_open_ip $uid $uuid $MONGODB_PORT`
-
-#
-# Fix permissions
-#
-observe_setup_app_home
-
-#
-# Source database connection variables
-#
-. $APP_HOME/.env/OPENSHIFT_MONGODB_DB_HOST
-. $APP_HOME/.env/OPENSHIFT_MONGODB_DB_PORT
-. $APP_HOME/.env/OPENSHIFT_MONGODB_DB_PASSWORD
-. $APP_HOME/.env/OPENSHIFT_MONGODB_DB_USERNAME
-
-ORIG_DB_HOST=$OPENSHIFT_MONGODB_DB_HOST
-
-#
-# Setup environment variables
-#
-echo "export OPENSHIFT_MONGODB_DB_USERNAME='$OPENSHIFT_MONGODB_DB_USERNAME'" > $APP_HOME/.env/OPENSHIFT_MONGODB_DB_USERNAME
-echo "export OPENSHIFT_MONGODB_DB_PASSWORD='$OPENSHIFT_MONGODB_DB_PASSWORD'" > $APP_HOME/.env/OPENSHIFT_MONGODB_DB_PASSWORD
-echo "export OPENSHIFT_MONGODB_DB_HOST='$MONGODB_IP'" > $APP_HOME/.env/OPENSHIFT_MONGODB_DB_HOST
-echo "export OPENSHIFT_MONGODB_DB_PORT='$MONGODB_PORT'" > $APP_HOME/.env/OPENSHIFT_MONGODB_DB_PORT
-echo "export OPENSHIFT_MONGODB_DB_URL='mongodb://$OPENSHIFT_MONGODB_DB_USERNAME:$OPENSHIFT_MONGODB_DB_PASSWORD@$MONGODB_IP:$MONGODB_PORT/'" > $APP_HOME/.env/OPENSHIFT_MONGODB_DB_URL
-
-#
-# Fix config file
-#
-if [ "$ORIG_DB_HOST" != "$MONGODB_IP" ]
-then
-    sed -i "s,$ORIG_DB_HOST,$MONGODB_IP," $MONGODB_DIR/etc/mongodb.conf
-    set_app_info "Connection URL: mongodb://$MONGODB_IP:$MONGODB_PORT/"
-fi
-
