--- cartridges/openshift-origin-cartridge-mongodb-2.2/info/hooks/post-move
@@ -1,27 +1,0 @@
-#!/bin/bash -e
-
-cartridge_type="mongodb-2.2"
-gear_type=$cartridge_type
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-CART_INFO_DIR=${CARTRIDGE_BASE_PATH}/embedded/$cartridge_type/info
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-
-setup_basic_vars
-
