--- cartridges/openshift-origin-cartridge-mongodb-2.2/info/hooks/pre-move
@@ -1,1 +1,0 @@
-#!/bin/bash
