--- cartridges/openshift-origin-cartridge-mysql-5.1/info/hooks/post-move
@@ -1,51 +1,0 @@
-#!/bin/bash
-
-# Cleanup app after moving to a new ip
-
-# Exit on any errors
-set -e
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/apache
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/network
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-
-setup_basic_vars
-cartridge_type="mysql-5.1"
-
-CART_INFO_DIR=${CARTRIDGE_BASE_PATH}/embedded/mysql-5.1/info
-source ${CART_INFO_DIR}/lib/util
-
-MYSQL_DIR="$APP_HOME/mysql-5.1/"
-
-. $APP_HOME/.env/OPENSHIFT_APP_NAME
-. $APP_HOME/.env/OPENSHIFT_MYSQL_DB_HOST
-IP=$OPENSHIFT_MYSQL_DB_HOST
-
-. $APP_HOME/.env/OPENSHIFT_MYSQL_DB_PORT
-. $APP_HOME/.env/OPENSHIFT_MYSQL_DB_PASSWORD
-. $APP_HOME/.env/OPENSHIFT_MYSQL_DB_USERNAME
-
-start_database
-
-    echo "
-DELETE FROM user WHERE Host<>'$IP' and User='$OPENSHIFT_MYSQL_DB_USERNAME';
-flush privileges;" | mysql -u $OPENSHIFT_MYSQL_DB_USERNAME -P $OPENSHIFT_MYSQL_DB_PORT -h $IP --password="$OPENSHIFT_MYSQL_DB_PASSWORD" -S "${MYSQL_DIR}/socket/mysql.sock" mysql > /dev/null || error "Failed to remove wildcard ip" 187
-
-stop_database
