--- cartridges/openshift-origin-cartridge-phpmyadmin-3.4/info/hooks/move
@@ -1,39 +1,0 @@
-#!/bin/bash
-
-# Move to a new ip
-
-# Exit on any errors
-set -e
-
-source "/etc/openshift/node.conf"
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/util
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/apache
-source ${CARTRIDGE_BASE_PATH}/abstract/info/lib/network
-
-while getopts 'd' OPTION
-do
-    case $OPTION in
-        d) set -x
-        ;;
-        ?) print_help
-        ;;
-    esac
-done
-
-[ $# -eq 3 ] || print_help
-
-namespace=`basename $2`
-application="$1"
-uuid=$3
-
-setup_basic_vars
-
-CART_INFO_DIR=${CARTRIDGE_BASE_PATH}/embedded/phpmyadmin-3.4/info
-
-PHPMYADMIN_DIR="$APP_HOME/phpmyadmin-3.4"
-observe_setup_var_lib_dir "$PHPMYADMIN_DIR"
-
-PHPMYADMIN_IP=`embedded_find_open_ip $uid $APP_HOME`
-echo "export OPENSHIFT_PHPMYADMIN_IP='$PHPMYADMIN_IP'" > $APP_HOME/.env/OPENSHIFT_PHPMYADMIN_IP
-$CART_INFO_DIR/bin/deploy_httpd_config.sh $application $uuid $PHPMYADMIN_IP
-$CART_INFO_DIR/bin/deploy_httpd_proxy.sh $application $namespace $uuid $PHPMYADMIN_IP
