--- cartridges/openshift-origin-cartridge-postgresql-8.4/info/hooks/post-move
@@ -1,1 +1,0 @@
-control-scripts
