--- cartridges/openshift-origin-cartridge-postgresql-8.4/info/hooks/pre-move
@@ -1,1 +1,0 @@
-#!/bin/bash
