--- cartridges/openshift-origin-cartridge-switchyard-0.6/info/hooks/move
@@ -1,1 +1,0 @@
-#!/bin/bash
