+++ plugins/msg-broker/mcollective/lib/mcollective_msg_broker_engine.rb
