+++ plugins/msg-broker/mcollective/lib/openshift-origin-msg-broker-mcollective.rb
@@ -2,7 +2,7 @@ require "openshift-origin-common"
 
 module OpenShift
   module McollectiveMsgBrokerModule
-    require 'mcollective_msg_brojer_engine' if defined?(Rails) && Rails::VERSION::MAJOR == 3
+    require 'mcollective_msg_broker_engine' if defined?(Rails) && Rails::VERSION::MAJOR == 3
   end
 end
