+++ plugins/msg-node/mcollective/src/openshift.rb
@@ -43,7 +43,7 @@ module MCollective
         Log.instance.info("cartridge_do_action validation = #{request[:cartridge]} #{request[:action]} #{request[:args]}")
         validate :cartridge, /\A[a-zA-Z0-9\.\-\/]+\z/
         validate :cartridge, :shellsafe
-        validate :action, /\A(app-create|app-destroy|env-var-add|env-var-remove|broker-auth-key-add|broker-auth-key-remove|authorized-ssh-key-add|authorized-ssh-key-remove|ssl-cert-add|ssl-cert-remove|configure|deconfigure|update-namespace|tidy|deploy-httpd-proxy|remove-httpd-proxy|restart-httpd-proxy|move|pre-move|post-move|info|post-install|post-remove|pre-install|reload|restart|start|status|stop|force-stop|add-alias|remove-alias|threaddump|cartridge-list|expose-port|frontend-backup|frontend-restore|frontend-create|frontend-destroy|frontend-update-name|frontend-update-namespace|frontend-connect|frontend-disconnect|frontend-connections|frontend-idle|frontend-unidle|frontend-check-idle|frontend-sts|frontend-no-sts|frontend-get-sts|aliases|ssl-cert-add|ssl-cert-remove|ssl-certs|frontend-to-hash|system-messages|connector-execute|get-quota|set-quota)\Z/
+        validate :action, /\A(app-create|app-destroy|env-var-add|env-var-remove|broker-auth-key-add|broker-auth-key-remove|authorized-ssh-key-add|authorized-ssh-key-remove|ssl-cert-add|ssl-cert-remove|configure|deconfigure|update-namespace|tidy|deploy-httpd-proxy|remove-httpd-proxy|restart-httpd-proxy|info|post-install|post-remove|pre-install|reload|restart|start|status|stop|force-stop|add-alias|remove-alias|threaddump|cartridge-list|expose-port|frontend-backup|frontend-restore|frontend-create|frontend-destroy|frontend-update-name|frontend-update-namespace|frontend-connect|frontend-disconnect|frontend-connections|frontend-idle|frontend-unidle|frontend-check-idle|frontend-sts|frontend-no-sts|frontend-get-sts|aliases|ssl-cert-add|ssl-cert-remove|ssl-certs|frontend-to-hash|system-messages|connector-execute|get-quota|set-quota)\Z/
         validate :action, :shellsafe
         cartridge                  = request[:cartridge]
         action                     = request[:action]
@@ -721,55 +721,6 @@ module MCollective
           return -1, e.message
         else
           return 0, ""
-        end
-      end
-
-      def oo_move(args)
-        cart_name = args['--cart-name']
-        idle      = args['--idle']
-
-        output = ""
-        begin
-          container = get_app_container_from_args(args)
-          output    = container.move(cart_name, idle)
-        rescue Exception => e
-          Log.instance.info e.message
-          Log.instance.info e.backtrace
-          return -1, e.message
-        else
-          return 0, output
-        end
-      end
-
-      def oo_pre_move(args)
-        cart_name = args['--cart-name']
-
-        output = ""
-        begin
-          container = get_app_container_from_args(args)
-          output    = container.pre_move(cart_name)
-        rescue Exception => e
-          Log.instance.info e.message
-          Log.instance.info e.backtrace
-          return -1, e.message
-        else
-          return 0, output
-        end
-      end
-
-      def oo_post_move(args)
-        cart_name = args['--cart-name']
-
-        output = ""
-        begin
-          container = get_app_container_from_args(args)
-          output    = container.post_move(cart_name)
-        rescue Exception => e
-          Log.instance.info e.message
-          Log.instance.info e.backtrace
-          return -1, e.message
-        else
-          return 0, output
         end
       end
